Add explicit integer-width casts and improve pointer arithmetic (#125)
 Add explicit width casts at sites where FPC's "Converting the operands to
Int64 before doing the add/subtract/multiply" and "Mixing signed expressions
and longwords gives a 64bit result" hints fire because a narrow (32-bit)
sub-operation feeds a wider expression. The casts make the intended width
explicit and silence the hints on both 32-bit and 64-bit; the computed values
are unchanged on every word size.

- ClpMlDsaCore.pas (NTT): Int64(LZeta) * (Int64(LA0) - Int64(LA1)) so the
  coefficient subtract widens before the multiply.
- ClpMlKemCore.pas (compress): (Int64(LC shl n) + (MlKemQ shr 8)) so the inner
  add runs in Int64; (decompress) cast the leading shift operand to UInt32 so
  the bit-packing OR no longer mixes a signed Int32 with a longword.
- ClpPoly1305.pas: (Int64(Int32(H4 shr 26)) - 1) * 5 so the subtract is Int64.
- ClpDateTimeUtilities.pas: Int64(LOffsetHours) * 60 + LOffsetMinutes for the
  UTC offset (IncMinute already takes Int64).

Deliberately left untouched: TMod.Inverse32, TX448Field.Sub, and the ML-DSA
eta=2 mod-5 reduction. Their hints are 32-bit-only false positives over
intentional fixed-width (UInt32) modular arithmetic; FPC flags any
UInt32*UInt32 / UInt32+UInt32 on a 32-bit target regardless of casts, and
widening to Int64 would be wrong for Inverse32 (it relies on mod-2^32
wraparound). These are left as the original code rather than suppressed.

FPC flags "Conversion between ordinals and pointers is not portable" at sites
that cast a pointer through NativeUInt. These are false positives (NativeUInt is
pointer-sized on all targets, so the conversions are size-safe on Delphi/FPC and
32/64-bit), but the sites that only need address arithmetic or comparison can
drop the integer round-trip entirely, since {$POINTERMATH ON} is global.

- ClpOcbBlockCipher.pas: compute the block-0 lookahead pointer with pointer
  arithmetic, PByte(@AInput[AInOff]) - BLOCK_SIZE, instead of
  Pointer(NativeUInt(@AInput[AInOff]) - BLOCK_SIZE). Removes the hint and the
  warning; same address.
- ClpAesEngineX86.pas (ProcessBlock / ProcessFourBlocks / ProcessEightBlocks):
  do the disjoint-vs-overlap test by comparing the PByte arguments directly
  (AOutput >= AInput, AOutput < AInput + N) and drop the now-unused
  LSrcAddr/LDstAddr NativeUInt locals. Pointer relational comparison is the same
  unsigned address comparison the NativeUInt casts performed.

Left untouched: the alignment fast-path tests in ClpBinaryPrimitives.Swap*,
ClpByteUtilities.Xor/XorTo, and TAesEngineX86.AllocAlignedKeys. These genuinely
need the address as an integer for `and`-masking, so their (harmless) hints
remain - no type alias, no suppression.

+++ CryptoLib/src/Crypto/Kems/MlKem/ClpMlKemCore.pas
@@ -740,7 +740,7 @@ begin
         for LK := 0 to 7 do
         begin
           LC := FVec[LI].Coeffs[8 * LJ + LK];
-          LT8[LK] := SmallInt((((Int64((LC shl 4) + (MlKemQ shr 8)) * 165141429) shr 32) and $7FF));
+          LT8[LK] := SmallInt(((((Int64(LC shl 4) + (MlKemQ shr 8)) * 165141429) shr 32) and $7FF));
         end;
         ABuf[LPos + 0] := Byte(LT8[0] shr 0);
         ABuf[LPos + 1] := Byte((LT8[0] shr 8) or (LT8[1] shl 3));
@@ -766,7 +766,7 @@ begin
         for LK := 0 to 3 do
         begin
           LC := FVec[LI].Coeffs[4 * LJ + LK];
-          LT4[LK] := SmallInt((((Int64((LC shl 3) + (MlKemQ shr 8)) * 165141429) shr 32) and $3FF));
+          LT4[LK] := SmallInt(((((Int64(LC shl 3) + (MlKemQ shr 8)) * 165141429) shr 32) and $3FF));
         end;
         ABuf[LPos + 0] := Byte(LT4[0] shr 0);
         ABuf[LPos + 1] := Byte((LT4[0] shr 8) or (LT4[1] shl 2));
@@ -805,14 +805,14 @@ begin
         LC9 := ABuf[LPos + 9];
         LC10 := ABuf[LPos + 10];
         System.Inc(LPos, 11);
-        LT0 := SmallInt((LC0 shr 0) or (UInt16(LC1) shl 8));
-        LT1 := SmallInt((LC1 shr 3) or (UInt16(LC2) shl 5));
-        LT2 := SmallInt((LC2 shr 6) or (UInt16(LC3) shl 2) or (UInt16(LC4 shl 10)));
-        LT3 := SmallInt((LC4 shr 1) or (UInt16(LC5) shl 7));
-        LT4 := SmallInt((LC5 shr 4) or (UInt16(LC6) shl 4));
-        LT5 := SmallInt((LC6 shr 7) or (UInt16(LC7) shl 1) or (UInt16(LC8 shl 9)));
-        LT6 := SmallInt((LC8 shr 2) or (UInt16(LC9) shl 6));
-        LT7 := SmallInt((LC9 shr 5) or (UInt16(LC10) shl 3));
+        LT0 := SmallInt((UInt32(LC0) shr 0) or (UInt16(LC1) shl 8));
+        LT1 := SmallInt((UInt32(LC1) shr 3) or (UInt16(LC2) shl 5));
+        LT2 := SmallInt((UInt32(LC2) shr 6) or (UInt16(LC3) shl 2) or (UInt16(LC4 shl 10)));
+        LT3 := SmallInt((UInt32(LC4) shr 1) or (UInt16(LC5) shl 7));
+        LT4 := SmallInt((UInt32(LC5) shr 4) or (UInt16(LC6) shl 4));
+        LT5 := SmallInt((UInt32(LC6) shr 7) or (UInt16(LC7) shl 1) or (UInt16(LC8 shl 9)));
+        LT6 := SmallInt((UInt32(LC8) shr 2) or (UInt16(LC9) shl 6));
+        LT7 := SmallInt((UInt32(LC9) shr 5) or (UInt16(LC10) shl 3));
         FVec[LI].Coeffs[LJ + 0] := SmallInt(((LT0 and $7FF) * MlKemQ + 1024) shr 11);
         FVec[LI].Coeffs[LJ + 1] := SmallInt(((LT1 and $7FF) * MlKemQ + 1024) shr 11);
         FVec[LI].Coeffs[LJ + 2] := SmallInt(((LT2 and $7FF) * MlKemQ + 1024) shr 11);
@@ -838,10 +838,10 @@ begin
         LC3 := ABuf[LPos + 3];
         LC4 := ABuf[LPos + 4];
         System.Inc(LPos, 5);
-        LT0 := SmallInt((LC0 shr 0) or (UInt16(LC1) shl 8));
-        LT1 := SmallInt((LC1 shr 2) or (UInt16(LC2) shl 6));
-        LT2 := SmallInt((LC2 shr 4) or (UInt16(LC3) shl 4));
-        LT3 := SmallInt((LC3 shr 6) or (UInt16(LC4) shl 2));
+        LT0 := SmallInt((UInt32(LC0) shr 0) or (UInt16(LC1) shl 8));
+        LT1 := SmallInt((UInt32(LC1) shr 2) or (UInt16(LC2) shl 6));
+        LT2 := SmallInt((UInt32(LC2) shr 4) or (UInt16(LC3) shl 4));
+        LT3 := SmallInt((UInt32(LC3) shr 6) or (UInt16(LC4) shl 2));
         FVec[LI].Coeffs[LJ + 0] := SmallInt(((LT0 and $3FF) * MlKemQ + 512) shr 10);
         FVec[LI].Coeffs[LJ + 1] := SmallInt(((LT1 and $3FF) * MlKemQ + 512) shr 10);
         FVec[LI].Coeffs[LJ + 2] := SmallInt(((LT2 and $3FF) * MlKemQ + 512) shr 10);
